backend/foodgram/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
asgiref==3.2.10
django-filter==21.1
Django==2.2.16
djangorestframework==3.12.4
gunicorn==20.0.4
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
pytest==6.2.4
python-dotenv==0.20.0
Pillow==9.4.0
djoser==2.1.0
psycopg2-binary==2.8.6
drf_extra_fields==3.4.1